Allow running directly or deploying as war
The change to the opencv version is required, as the previous version
had a bug in the library loading logic. Going to the next minor version
with that bug fixed isn't an option, as currently the docker image with
OCR4all and LAREX only works, if the native library versions differ (so
LAREX OpenCV version must be different from OCR4all).